requirements.txt
streamlit>=1.20.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.7.5
langchain>=0.0.200
openai>=0.27.0
requests>=2.28.0
pillow>=9.0.0
tqdm>=4.64.0
